requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.7
async-generator==1.10
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bleach==3.3.1
certifi==2020.4.5.1
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
colorama==0.4.4
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
Django==3.0.5
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0
entrypoints==0.3
filelock==3.0.12
future==0.18.2
idna==2.9
importlib-metadata==1.6.0
iniconfig==1.1.1
install==1.3.4
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
more-itertools==8.2.0
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
packaging==20.3
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
platformdirs==2.1.0
pluggy==0.13.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
py==1.8.1
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==5.4.1
pytest-django==3.9.0
python-dateutil==2.8.2
pytz==2019.3
pywin32==301
pywinpty==1.1.3
pyzmq==22.1.0
requests==2.23.0
Send2Trash==1.7.1
six==1.14.0
sqlparse==0.3.1
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.25.9
virtualenv==20.6.0
wcwidth==0.1.9
webencodings==0.5.1
zipp==3.1.0
